requirements.txt
Babel==2.2.0
Django==1.9.1
django-allauth==0.24.1
django-crispy-forms==1.5.2
django-phonenumber-field==1.0.0
oauthlib==1.0.3
phonenumbers==7.2.2
Pillow==3.1.0
python-openid==2.2.5
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
wheel==0.24.0
